feat: implement scroll-driven hero component
